Changed package name to feasts
First suggested by @robjhyndman as an anagram for "FEatures And Statistics for Time Series".

Meaning slightly changed by myself to be "Feature Extraction And Statistics for Time Series".

This name works well with the theme of tidy modelling packages, and is meaningful in its own right. The package feasts on time series data to produce summaries/statistics/features/plots. It can also be thought to feast on the delicious models produced with the recipes package :smile: :fork_and_knife:

@robjhyndman @earowang @dicook